structuring code
