+++ .env.example
@@ -1,0 +1,2 @@
+AUTH_EMAIL=
+AUTH_PASS=
